refactor(BA-5638): introduce UserSystemRoleSpec instead of extending auth UserData
Create a dedicated frozen dataclass that implements ScopeSystemRoleData,
taking only a user_id. This avoids polluting auth UserData with RBAC
protocol methods.
