+++ src/ai/backend/manager/data/auth/types.py
@@ -1,15 +1,7 @@
 import uuid
-from collections.abc import Iterable, Mapping
 from dataclasses import dataclass
 from datetime import datetime
 
-from ai.backend.common.data.permission.types import RBACElementType
-from ai.backend.manager.data.permission.id import ScopeId
-from ai.backend.manager.data.permission.types import (
-    EntityType,
-    OperationType,
-    ScopeType,
-)
 from ai.backend.manager.models.user import UserRole, UserStatus
 
 
@@ -49,23 +41,6 @@ class UserData:
     integration_name: str | None
     resource_policy: str
     sudo_session_enabled: bool
-
-    def scope_id(self) -> ScopeId:
-        return ScopeId(
-            scope_type=ScopeType.USER,
-            scope_id=str(self.uuid),
-        )
-
-    def role_name(self) -> str:
-        return f"user-{str(self.uuid)[:8]}"
-
-    def entity_operations(self) -> Mapping[RBACElementType, Iterable[OperationType]]:
-        resource_entity_permissions = {
-            entity.to_element(): OperationType.owner_operations()
-            for entity in EntityType.owner_accessible_entity_types_in_user()
-        }
-        user_permissions = OperationType.owner_operations() - {OperationType.CREATE}
-        return {RBACElementType.USER: user_permissions, **resource_entity_permissions}
 
 
 @dataclass
